Add support for BigDecimal in account management and SQL queries
- Introduced the `bigdecimal` crate to handle precise decimal arithmetic for account balances and transaction amounts.
- Updated the `AccountService` struct and related methods to use `BigDecimal` instead of `u64` for monetary values, enhancing accuracy in financial calculations.
- Modified SQL migration scripts to change column types from `BIGINT` to `NUMERIC`, ensuring compatibility with the new data type.
- Added new SQL queries for retrieving account information with `BigDecimal` fields, improving data handling in the application.
- Created new JSON query files for SQL operations, reflecting the updated schema and ensuring proper integration with PostgreSQL.

+++ openmatching-account/migrations/20250918125000_init.up.sql
@@ -4,9 +4,9 @@
 CREATE TABLE accounts(
     username VARCHAR(255) NOT NULL,
     asset VARCHAR(255) NOT NULL,
-    balance BIGINT NOT NULL CHECK (balance >= 0),
-    pending_credit BIGINT NOT NULL CHECK (pending_credit >= 0),
-    pending_debit BIGINT NOT NULL CHECK (pending_debit >= 0),
+    balance NUMERIC NOT NULL CHECK (balance >= 0),
+    pending_credit NUMERIC NOT NULL CHECK (pending_credit >= 0),
+    pending_debit NUMERIC NOT NULL CHECK (pending_debit >= 0),
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     PRIMARY KEY (username, asset)
@@ -21,7 +21,7 @@ CREATE TABLE transactions(
     debitor VARCHAR(255) NOT NULL,
     creditor VARCHAR(255) NOT NULL,
     asset VARCHAR(255) NOT NULL,
-    amount BIGINT NOT NULL CHECK (amount > 0),
+    amount NUMERIC NOT NULL CHECK (amount > 0),
     status transaction_status NOT NULL,
     extra JSONB,
     external_id VARCHAR(255) UNIQUE,
